I made a good first cut at getting CKit to 64-bit. There's a few things left to do, but these are from the Solaris 64-bit port I was working on.

+++ src/muParserBase.h
@@ -298,7 +298,7 @@ class ParserBase
 		*/
 		mutable ParseFunction m_pParseFormula;
 		mutable int	m_iPos;	               ///< Global index within bytecode.
-		mutable const int *m_pCmdCode;     ///< Formula converted to bytecode, points to the data of the bytecode class.
+		mutable const long *m_pCmdCode;     ///< Formula converted to bytecode, points to the data of the bytecode class.
 		mutable bytecode_type m_vByteCode; ///< The Bytecode class.
 		mutable varmap_type m_UsedVar;     ///< Map holding the variable found in the current expression.
